@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Api::V1::RecommendationsController do
+describe Api::V1::RecommendationsController, :type => :controller  do
   render_views
 
   before :each do
@@ -25,8 +25,9 @@ describe Api::V1::RecommendationsController do
 
       get :index, :article_id => @recommendation.article_id
       response.should be_success
+      @load_content = :user
       rendered = Rabl.render(
-          [@recommendation], 'api/v1/recommendations/index', :view_path => 'app/views')
+          [@recommendation], 'api/v1/recommendations/index', :view_path => 'app/views', scope: self)
       response.body.should eq(rendered)
     end
   end
@@ -39,21 +40,24 @@ describe Api::V1::RecommendationsController do
 
       get :index, :user_id => @recommendation.user_id
       response.should be_success
+      @load_content = :article
       rendered = Rabl.render(
-          [@recommendation], 'api/v1/recommendations/index', :view_path => 'app/views')
+          [@recommendation], 'api/v1/recommendations/index', :view_path => 'app/views', scope: self)
       response.body.should eq(rendered)
     end
   end
 
   describe "POST 'articles/:article_id/recommendations'" do
     it "should error if the user already recommended the article" do
-      request.env['HTTP_AUTHORIZATION'] = %Q{Token token="#{@recommendation.user.authentication_token}"}
+      auth_headers = @recommendation.user.create_new_auth_token
+      request.headers.merge!(auth_headers)
       post :create, :article_id => @recommendation.article_id
       response.code.should eq('422')
     end
 
     it "should create an article recommendation of the current user" do
-      request.env['HTTP_AUTHORIZATION'] = %Q{Token token="#{@user.authentication_token}"}
+      auth_headers = @user.create_new_auth_token
+      request.headers.merge!(auth_headers)
       post :create, :article_id => @recommendation.article_id
       response.should be_success
       parsed_response = JSON.parse(response.body)
@@ -63,7 +67,6 @@ describe Api::V1::RecommendationsController do
     end
 
     it "should return 401 when the user is not logged in or unauthorized" do
-      request.env['HTTP_AUTHORIZATION'] = nil
       post :create, :article_id => @recommendation.article_id
       response.code.should eq('401')
     end
@@ -72,13 +75,15 @@ describe Api::V1::RecommendationsController do
   describe "DELETE 'articles/:article_id/recommendations/:id'" do
 
     it "should return 401 when the user is not logged in or unauthorized" do
-      request.env['HTTP_AUTHORIZATION'] = %Q{Token token="#{@user.authentication_token}"}
+      auth_headers = @user.create_new_auth_token
+      request.headers.merge!(auth_headers)
       delete :destroy, :article_id => @recommendation.article_id, :id => @recommendation.id
       response.code.should eq('401')
     end
 
     it "should delete an article recommendation of the current user" do
-      request.env['HTTP_AUTHORIZATION'] = %Q{Token token="#{@recommendation.user.authentication_token}"}
+      auth_headers = @recommendation.user.create_new_auth_token
+      request.headers.merge!(auth_headers)
       delete :destroy, :article_id => @recommendation.article_id, :id => @recommendation.id
       response.should be_success
 
