@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Api::V1::UsersController do
+describe Api::V1::UsersController, :type => :controller  do
   render_views
 
   before (:each) do
@@ -9,7 +9,7 @@ describe Api::V1::UsersController do
 
   describe 'GET users' do
     it 'should return only publishers' do
-      get :index
+      get :index, format: :json
       response.should be_success
       rendered = Rabl.render(
           [], 'api/v1/users/index', :view_path => 'app/views')
@@ -19,7 +19,7 @@ describe Api::V1::UsersController do
 
   describe 'GET users/:id' do
     it 'should return the user' do
-      get :show, :id => @user.id
+      get :show, :id => @user.id, format: :json
       response.should be_success
       rendered = Rabl.render(
           @user, 'api/v1/users/show', :view_path => 'app/views')
