@@ -1,29 +1,46 @@
 class Article < ActiveRecord::Base
-  include Utils
+  include Concerns::Utils
+  paginates_per 15
 
-  before_save :published_post, :time_to_read
-  after_save :update_published_articles_count
+  belongs_to :user
+  belongs_to :topic
+  belongs_to :category
+  has_many :recommendations, dependent: :destroy
+  has_many :comments, dependent: :destroy
+
+  scope :publishings, -> { where(published: true) }
+  scope :drafts, -> { where(published: false) }
 
   scope :popular, -> { order('hotness DESC') }
   scope :best, -> { order('recommendations_count DESC') }
   scope :recents, -> { order('published_at DESC') }
 
-  belongs_to :user
-  has_many :recommendations, :dependent => :destroy
-  has_many :comments, :dependent => :destroy
-  belongs_to :topic
-  has_one :category, through: :topic, autosave: false
+  before_save :published_post, :time_to_read
+  after_save :update_published_articles_count
 
   dragonfly_accessor :cover do
     storage_options do |attachment|
       { headers: {"x-amz-acl" => "public-read-write"} }
     end
   end
-
   abs_url_for :cover
 
-  scope :public, -> { where(published: true) }
-  scope :drafts, -> { where(published: false) }
+
+  def user_name
+    user.name
+  end
+  persistize :user_name
+
+  def topic_title
+    topic ? topic.title : nil
+  end
+  persistize :topic_title
+
+  def category_color
+    category ? category.color : nil
+  end
+  persistize :category_color
+
 
   # Instance Methods.
   def publish!
@@ -43,9 +60,10 @@ class Article < ActiveRecord::Base
   end
 
   def next
-    next_article = Article.public.popular.where('hotness < ?', hotness).first
+    query = Article.publishings.popular
+    next_article = query.where('hotness < ?', hotness).first
     if next_article.nil?
-        next_article = Article.public.popular.where('hotness > ?', hotness).first
+      next_article = query.where('hotness > ?', hotness).first
     end
     next_article
   end
@@ -60,6 +78,7 @@ class Article < ActiveRecord::Base
 
   def update_published_articles_count
     self.user.published_articles_count = self.user.published_articles.count
+    self.user.save
   end
 
 end
