@@ -1,6 +1,5 @@
 class ArticleRankingWorker
   include Sidekiq::Worker
-
   sidekiq_options retry: false
 
   app_start_time = Time.local(2014, 1, 1)
@@ -8,7 +7,7 @@ class ArticleRankingWorker
   @@RANKING_LOG_BASE = (ENV['RANKING_LOG_BASE'] || 2).to_i
   @@RANKING_TIME_SCALE_CONSTANT = (ENV['RANKING_TIME_SCALE_CONSTANT']|| 45000).to_i
   @@RANKING_RECOMMENDATION_WEIGHT = (ENV['RANKING_RECOMMENDATION_WEIGHT'] || 2).to_i
-  @@RANKING_COMMENTS_WEIGHT = (ENV['RANKING_COMMENTS_WEIGHT'].to_i || 1).to_i
+  @@RANKING_COMMENTS_WEIGHT = (ENV['RANKING_COMMENTS_WEIGHT'] || 1).to_i
 
   def self.score(recommendations_count, comments_count)
     (recommendations_count * @@RANKING_RECOMMENDATION_WEIGHT +
@@ -34,14 +33,7 @@ class ArticleRankingWorker
       puts "Updating ranking for #{article.id} = #{hotness}..."
       article.update_attributes(hotness: hotness)
     else
-      Article.find_each do |article|
-        hotness = ArticleRankingWorker.hot(
-            article.recommendations_count,
-            article.comments_count,
-            article.published_at)
-        puts "Updating ranking for #{article.id} = #{hotness}..."
-        article.update_attributes(hotness: hotness)
-      end
+      puts 'article_id was not passed. Nothing to do.'
     end
 
     puts 'Ranking articles worker finished!'
