@@ -1,11 +1,12 @@
-Backend::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
+  config.i18n.default_locale = :ar
 
   # Code is not reloaded between requests.
   config.cache_classes = true
 
   # Eager load code on boot. This eager loads most of Rails and
-  # your application in memory, allowing both thread web servers
+  # your application in memory, allowing both threaded web servers
   # and those relying on copy on write to perform better.
   # Rake tasks automatically ignore this option for performance.
   config.eager_load = true
@@ -32,8 +33,7 @@ Backend::Application.configure do
   # Generate digests for assets URLs.
   config.assets.digest = true
 
-  # Version of your assets, change this if you want to expire all your assets.
-  config.assets.version = '1.0'
+  # `config.assets.precompile` and `config.assets.version` have moved to config/initializers/assets.rb
 
   # Specifies the header that your server uses for sending files.
   # config.action_dispatch.x_sendfile_header = "X-Sendfile" # for apache
@@ -57,16 +57,12 @@ Backend::Application.configure do
   # Enable serving of images, stylesheets, and JavaScripts from an asset server.
   # config.action_controller.asset_host = "http://assets.example.com"
 
-  # Precompile additional assets.
-  # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
-  # config.assets.precompile += %w( search.js )
-
   # Ignore bad email addresses and do not raise email delivery errors.
   # Set this to true and configure the email server for immediate delivery to raise delivery errors.
   # config.action_mailer.raise_delivery_errors = false
 
   # Enable locale fallbacks for I18n (makes lookups for any locale fall back to
-  # the I18n.default_locale when a translation can not be found).
+  # the I18n.default_locale when a translation cannot be found).
   config.i18n.fallbacks = true
 
   # Send deprecation notices to registered listeners.
@@ -80,27 +76,35 @@ Backend::Application.configure do
   # Use default logging formatter so that PID and timestamp are not suppressed.
   config.log_formatter = ::Logger::Formatter.new
 
+  # Do not dump schema after migrations.
+  config.active_record.dump_schema_after_migration = false
+
+  OmniAuth.config.full_host = "http://#{ENV['API_HOST']}"
+
+  # Notification emails configurations.
+  config.roadie.url_options = { host: ENV['WEB_CLIENT_HOST'], scheme: "http" }
+  config.action_mailer.default_options = { from: ENV['DEFAULT_FROM_EMAIL'] }
   config.action_mailer.default_url_options = { :host => ENV['API_HOST'] }
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.smtp_settings = {
-    :enable_starttls_auto => true,
-    :address => "smtp.gmail.com",
+    :authentication => :plain,
+    :address => ENV['SMTP_HOSTNAME'],
     :port => 587,
-    :domain => "manshar.me",
-    :authentication => 'login',
-    :user_name => ENV['GMAIL_USERNAME'],
-    :password => ENV['GMAIL_PWD']
+    :domain => ENV['SMTP_DOMAIN'],
+    :user_name => ENV['SMTP_USERNAME'],
+    :password => ENV['SMTP_PASSWORD'],
   }
-  
+
+
   # Allow Cross-Origin Resource Sharing header to allow cross
   # domain xhr requests.
   config.middleware.insert_before Warden::Manager, Rack::Cors do
     allow do
-      origins ENV['WEB_CLIENT_HOST']
+      origins ENV['CORS_ORIGINS'] || ENV['WEB_CLIENT_HOST']
       resource '*',
       :headers => :any,
+      :expose => ['access-token', 'expiry', 'token-type', 'uid', 'client'],
       :methods => [:get, :delete, :post, :put, :options]
     end
   end
-
 end
