@@ -10,10 +10,10 @@ Devise.setup do |config|
   # Configure the e-mail address which will be shown in Devise::Mailer,
   # note that it will be overwritten if you use your own mailer class
   # with default "from" parameter.
-  config.mailer_sender = 'info@manshar.me'
+  config.mailer_sender = ENV['DEFAULT_FROM_EMAIL']
 
   # Configure the class responsible to send e-mails.
-  # config.mailer = 'Devise::Mailer'
+  config.mailer = 'AccountsMailer'
 
   # ==> ORM configuration
   # Load and configure the ORM. Supports :active_record (default) and
@@ -106,7 +106,7 @@ Devise.setup do |config|
   # able to access the website for two days without confirming his account,
   # access will be blocked just in the third day. Default is 0.days, meaning
   # the user cannot access the website without confirming his account.
-  config.allow_unconfirmed_access_for = 2.days
+  config.allow_unconfirmed_access_for = 5.days
 
   # A period that the user is allowed to confirm their account before their
   # token becomes invalid. For example, if set to 3.days, the user can confirm
@@ -196,10 +196,6 @@ Devise.setup do |config|
   # Require the `devise-encryptable` gem when using anything other than bcrypt
   # config.encryptor = :sha512
 
-  # ==> Configuration for :token_authenticatable
-  # Defines name of the authentication token params key
-  config.token_authentication_key = :auth_token
-
   # ==> Scopes configuration
   # Turn scoped views on. Before rendering "sessions/new", it will first check for
   # "users/sessions/new". It's turned off by default because it's slower if you
@@ -245,41 +241,6 @@ Devise.setup do |config|
   # Tell warden to use our own custom failure app.
   config.warden do |manager|
     manager.failure_app = UnauthorizedResponder
-  end
-
-
-  # ==> Mountable engine configurations
-  # When using Devise inside an engine, let's call it `MyEngine`, and this engine
-  # is mountable, there are some extra configurations to be taken into account.
-  # The following options are available, assuming the engine is mounted as:
-  #
-  #     mount MyEngine, at: '/my_engine'
-  #
-  # The router that invoked `devise_for`, in the example above, would be:
-  # config.router_name = :my_engine
-  #
-  # When using omniauth, Devise cannot automatically set Omniauth path,
-  # so you need to do it manually. For the users scope, it would be:
-  # config.omniauth_path_prefix = '/my_engine/users/auth'
-
-  # This sets up devise to use the auth_token from the HTTP_AUTHORIZATION header.
-  # Ref: http://goo.gl/6VDFVn
-  require 'devise/strategies/token_authenticatable'
-  module Devise
-    module Strategies
-      class TokenAuthenticatable < Authenticatable
-        def params_auth_hash
-          return_params = if params[scope].kind_of?(Hash) && params[scope].has_key?(authentication_keys.first)
-            params[scope]
-          else
-            params
-          end
-          token = ActionController::HttpAuthentication::Token.token_and_options(request)
-          return_params.merge!(:auth_token => token[0]) if token
-          return_params
-        end
-      end
-    end
   end
 
 end
