@@ -1,21 +1,22 @@
 Mailboxer.setup do |config|
 
-  #Configures if you application uses or not email sending for Notifications and Messages
-  config.uses_emails = true
+  # Configures if your application uses or not email sending for Notifications
+  # and Messages.
+  config.uses_emails = false
 
-  #Configures the default from for emails sent for Messages and Notifications
-  config.default_from = "noreply@manshar.com"
+  # Configures the default from for emails sent for Messages and Notifications.
+  config.default_from = ENV['DEFAULT_FROM_EMAIL']
 
-  #Configures the methods needed by mailboxer
+  # Configures the methods needed by mailboxer.
   config.email_method = :mailboxer_email
   config.name_method = :name
 
-  #Configures if you use or not a search engine and which one you are using
-  #Supported engines: [:solr,:sphinx]
+  # Configures if you use or not a search engine and which one you are using.
+  # Supported engines: [:solr,:sphinx].
   config.search_enabled = false
   config.search_engine = :solr
 
-  #Configures maximum length of the message subject and body
+  # Configures maximum length of the message subject and body.
   config.subject_max_length = 255
   config.body_max_length = 32000
 end
