@@ -1,7 +1,7 @@
 class UnauthorizedResponder < Devise::FailureApp
   def respond
-    self.status = 401 
+    self.status = 401
     self.content_type = 'json'
     self.response_body = {"errors" => "Unauthorized."}.to_json
-  end 
+  end
 end
