@@ -1,11 +1,12 @@
 # Be sure to restart your server when you modify this file.
-#
-# This file contains settings for ActionController::ParamsWrapper
 
-# Enable parameter wrapping for JSON.
-# ActiveSupport.on_load(:action_controller) do
-#   wrap_parameters format: [:json] if respond_to?(:wrap_parameters)
-# end
+# This file contains settings for ActionController::ParamsWrapper which
+# is enabled by default.
+
+# Enable parameter wrapping for JSON. You can disable this by setting :format to an empty array.
+ActiveSupport.on_load(:action_controller) do
+  wrap_parameters format: [:json] if respond_to?(:wrap_parameters)
+end
 
 # To enable root element in JSON for ActiveRecord objects.
 # ActiveSupport.on_load(:active_record) do
