@@ -11,12 +11,12 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150203055410) do
+ActiveRecord::Schema.define(version: 20150215055322) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "articles", force: true do |t|
+  create_table "articles", force: :cascade do |t|
     t.string   "title",                 default: "",    null: false
     t.string   "tagline",               default: "",    null: false
     t.text     "body",                  default: "",    null: false
@@ -32,9 +32,13 @@ ActiveRecord::Schema.define(version: 20150203055410) do
     t.datetime "published_at"
     t.integer  "reading_time"
     t.integer  "topic_id"
+    t.string   "user_name"
+    t.string   "topic_title"
+    t.string   "category_color"
+    t.integer  "category_id"
   end
 
-  create_table "categories", force: true do |t|
+  create_table "categories", force: :cascade do |t|
     t.string   "title",                              null: false
     t.text     "description"
     t.string   "image_uid"
@@ -47,7 +51,7 @@ ActiveRecord::Schema.define(version: 20150203055410) do
     t.string   "icon_cssclass",  default: "fa-tag"
   end
 
-  create_table "comments", force: true do |t|
+  create_table "comments", force: :cascade do |t|
     t.string   "guid"
     t.text     "body",       null: false
     t.integer  "article_id", null: false
@@ -56,7 +60,7 @@ ActiveRecord::Schema.define(version: 20150203055410) do
     t.datetime "updated_at"
   end
 
-  create_table "images", force: true do |t|
+  create_table "images", force: :cascade do |t|
     t.string   "title",      default: "", null: false
     t.text     "caption",    default: "", null: false
     t.integer  "user_id",                 null: false
@@ -66,7 +70,15 @@ ActiveRecord::Schema.define(version: 20150203055410) do
     t.datetime "updated_at"
   end
 
-  create_table "mailboxer_conversation_opt_outs", force: true do |t|
+  create_table "links", force: :cascade do |t|
+    t.string   "title",      default: "", null: false
+    t.string   "url",        default: "", null: false
+    t.integer  "user_id",                 null: false
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
+
+  create_table "mailboxer_conversation_opt_outs", force: :cascade do |t|
     t.integer "unsubscriber_id"
     t.string  "unsubscriber_type"
     t.integer "conversation_id"
@@ -75,13 +87,13 @@ ActiveRecord::Schema.define(version: 20150203055410) do
   add_index "mailboxer_conversation_opt_outs", ["conversation_id"], name: "index_mailboxer_conversation_opt_outs_on_conversation_id", using: :btree
   add_index "mailboxer_conversation_opt_outs", ["unsubscriber_id", "unsubscriber_type"], name: "index_mailboxer_conversation_opt_outs_on_unsubscriber_id_type", using: :btree
 
-  create_table "mailboxer_conversations", force: true do |t|
+  create_table "mailboxer_conversations", force: :cascade do |t|
     t.string   "subject",    default: ""
     t.datetime "created_at",              null: false
     t.datetime "updated_at",              null: false
   end
 
-  create_table "mailboxer_notifications", force: true do |t|
+  create_table "mailboxer_notifications", force: :cascade do |t|
     t.string   "type"
     t.text     "body"
     t.string   "subject",              default: ""
@@ -104,7 +116,7 @@ ActiveRecord::Schema.define(version: 20150203055410) do
   add_index "mailboxer_notifications", ["sender_id", "sender_type"], name: "index_mailboxer_notifications_on_sender_id_and_sender_type", using: :btree
   add_index "mailboxer_notifications", ["type"], name: "index_mailboxer_notifications_on_type", using: :btree
 
-  create_table "mailboxer_receipts", force: true do |t|
+  create_table "mailboxer_receipts", force: :cascade do |t|
     t.integer  "receiver_id"
     t.string   "receiver_type"
     t.integer  "notification_id",                            null: false
@@ -119,7 +131,7 @@ ActiveRecord::Schema.define(version: 20150203055410) do
   add_index "mailboxer_receipts", ["notification_id"], name: "index_mailboxer_receipts_on_notification_id", using: :btree
   add_index "mailboxer_receipts", ["receiver_id", "receiver_type"], name: "index_mailboxer_receipts_on_receiver_id_and_receiver_type", using: :btree
 
-  create_table "recommendations", force: true do |t|
+  create_table "recommendations", force: :cascade do |t|
     t.integer  "article_id"
     t.integer  "user_id"
     t.datetime "created_at"
@@ -128,14 +140,14 @@ ActiveRecord::Schema.define(version: 20150203055410) do
 
   add_index "recommendations", ["user_id", "article_id"], name: "index_recommendations_on_user_id_and_article_id", unique: true, using: :btree
 
-  create_table "thumbs", force: true do |t|
+  create_table "thumbs", force: :cascade do |t|
     t.string   "signature"
     t.string   "uid"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "topics", force: true do |t|
+  create_table "topics", force: :cascade do |t|
     t.string   "title",                      null: false
     t.integer  "articles_count", default: 0
     t.integer  "category_id",                null: false
@@ -143,7 +155,7 @@ ActiveRecord::Schema.define(version: 20150203055410) do
     t.datetime "updated_at"
   end
 
-  create_table "users", force: true do |t|
+  create_table "users", force: :cascade do |t|
     t.string   "email",                    default: "",     null: false
     t.string   "encrypted_password",       default: "",     null: false
     t.string   "reset_password_token"
@@ -163,23 +175,26 @@ ActiveRecord::Schema.define(version: 20150203055410) do
     t.datetime "updated_at"
     t.string   "name",                     default: "",     null: false
     t.string   "bio",                      default: ""
-    t.string   "avatar_uid",                                null: false
+    t.string   "avatar_uid"
     t.string   "avatar_name"
     t.integer  "recommendations_count",    default: 0
     t.integer  "comments_count",           default: 0
     t.string   "role",                     default: "user"
     t.integer  "published_articles_count", default: 0
+    t.string   "provider"
+    t.string   "uid",                      default: "",     null: false
+    t.text     "tokens"
+    t.string   "nickname"
   end
 
   add_index "users", ["authentication_token"], name: "index_users_on_authentication_token", unique: true, using: :btree
   add_index "users", ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true, using: :btree
   add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
+  add_index "users", ["nickname"], name: "index_users_on_nickname", unique: true, using: :btree
   add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
+  add_index "users", ["uid"], name: "index_users_on_uid", unique: true, using: :btree
 
-  add_foreign_key "mailboxer_conversation_opt_outs", "mailboxer_conversations", name: "mb_opt_outs_on_conversations_id", column: "conversation_id"
-
-  add_foreign_key "mailboxer_notifications", "mailboxer_conversations", name: "notifications_on_conversation_id", column: "conversation_id"
-
-  add_foreign_key "mailboxer_receipts", "mailboxer_notifications", name: "receipts_on_notification_id", column: "notification_id"
-
+  add_foreign_key "mailboxer_conversation_opt_outs", "mailboxer_conversations", column: "conversation_id", name: "mb_opt_outs_on_conversations_id"
+  add_foreign_key "mailboxer_notifications", "mailboxer_conversations", column: "conversation_id", name: "notifications_on_conversation_id"
+  add_foreign_key "mailboxer_receipts", "mailboxer_notifications", column: "notification_id", name: "receipts_on_notification_id"
 end
