@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Api::V1::UsersArticlesController do
+describe Api::V1::UsersArticlesController, :type => :controller  do
   render_views
 
   before (:each) do
@@ -34,13 +34,16 @@ describe Api::V1::UsersArticlesController do
       get :index
       response.code.should eq('404')
 
-      request.env['HTTP_AUTHORIZATION'] = %Q{Token token="#{@article.user.authentication_token}"}
+      auth_headers = @article.user.create_new_auth_token
+      request.headers.merge!(auth_headers)
       get :index
       response.should be_success
       response.body.should eq([].to_json)
 
       @article.publish!
+      @article = Article.find(@article.id)
       @article.reload
+      get :index
 
       get :index
       response.should be_success
@@ -55,14 +58,16 @@ describe Api::V1::UsersArticlesController do
       get :drafts
       response.code.should eq('401')
 
-      request.env['HTTP_AUTHORIZATION'] = %Q{Token token="#{@user.authentication_token}"}
+      auth_headers = @user.create_new_auth_token
+      request.headers.merge!(auth_headers)
       get :drafts
       response.should be_success
       response.body.should eq([].to_json)
     end
 
     it 'should return all drafts for the current user' do
-      request.env['HTTP_AUTHORIZATION'] = %Q{Token token="#{@article.user.authentication_token}"}
+      auth_headers = @article.user.create_new_auth_token
+      request.headers.merge!(auth_headers)
       get :drafts
       response.should be_success
       rendered = Rabl.render(
