@@ -1,8 +1,9 @@
 'use strict';
 
 angular.module('webClientApp')
-  .controller('CategoryCtrl', ['$scope', '$rootScope', '$routeParams', '$location', '$filter', 'Category', 'Topic', 'CategoryArticle',
-      function ($scope, $rootScope, $routeParams, $location, $filter, Category, Topic, CategoryArticle) {
+  .controller('CategoryCtrl', ['$scope', '$rootScope', '$routeParams', '$location', '$filter', '$anchorScroll', 'Category', 'Topic', 'CategoryArticle',
+      function ($scope, $rootScope, $routeParams, $location, $filter, $anchorScroll, Category, Topic, CategoryArticle) {
+    $anchorScroll();
 
     Category.get({'categoryId': $routeParams.categoryId},
         function(resource) {
@@ -15,6 +16,8 @@ angular.module('webClientApp')
       $scope.category = resource;
     });
 
+    $rootScope.forceBar = true;
+
     // Get all topics in this category.
     $scope.topics = Topic.query({'categoryId': $routeParams.categoryId});
 
@@ -25,4 +28,27 @@ angular.module('webClientApp')
         function(articles) {
       $scope.articles = articles;
     });
+
+    $scope.showCategoriesPicker = function() {
+      $rootScope.$emit('openTopicPicker', {pickOnlyCategory: true});
+    };
+
+    var categorySelectedUnbind = $rootScope.$on('categorySelected',
+        function(event, data) {
+      $location.path('/categories/' + data.category.id);
+    });
+
+    $scope.getCardColor = function(color) {
+      return $filter('darker')(color, -0.4);
+    };
+
+    /**
+     * Make sure to cleanup the binded events and intervals when the user
+     * leaves to another controller.
+     */
+    var onDestroy = function () {
+      categorySelectedUnbind();
+    };
+    $scope.$on('$destroy', onDestroy);
+
   }]);
