@@ -12,9 +12,29 @@ angular.module('webClientApp')
     $scope.activeTab = $scope.order;
     Article.query({'order': $scope.order}, function(articles) {
       $scope.articles = articles;
-
+      $scope.hasNext = true;
       $scope.publishers = User.query();
     });
+
+    $rootScope.forceBar = false;
+
+    // TODO(mkhatib): Refactor this to move it to its own directive for listing
+    // articles with specific
+    var page = 1;
+    $scope.hasNext = false;
+    $scope.loadMoreArticles = function() {
+      $scope.inProgress = 'load-more';
+      Article.query({
+        'order': $scope.order,
+        'page': ++page
+      }, function(articles) {
+        if (!articles || !articles.length) {
+          $scope.hasNext = false;
+        }
+        Array.prototype.push.apply($scope.articles, articles);
+        $scope.inProgress = null;
+      });
+    };
 
     $scope.newArticle = function () {
       $location.path('/articles/new');
@@ -25,12 +45,14 @@ angular.module('webClientApp')
     };
 
     $scope.orderArticles = function (order) {
+      page = 1;
       $scope.activeTab = order;
       $scope.articles = [{ loading: true }, { loading: true },
           { loading: true }];
       $scope.order = order;
       Article.query({'order': order}, function(articles) {
         $scope.articles = articles;
+        $scope.hasNext = true;
       });
     };
 
@@ -46,6 +68,10 @@ angular.module('webClientApp')
         function(event, data) {
       $location.path('/categories/' + data.category.id);
     });
+
+    $scope.getCardColor = function(color) {
+      return color || '#C0C0C0';
+    };
 
     /**
      * Make sure to cleanup the binded events and intervals when the user
