@@ -1,8 +1,8 @@
 'use strict';
 
 angular.module('webClientApp')
-  .controller('ProfileCtrl', ['$scope', '$rootScope', '$location', '$routeParams', 'UserArticle', 'UserDraft', 'User','$analytics', '$window', 'Article', 'UserRecommendation', 'UserComment',
-    function ($scope, $rootScope, $location, $routeParams, UserArticle, UserDraft, User, $analytics, $window, Article, UserRecommendation, UserComment) {
+  .controller('ProfileCtrl', ['$scope', '$rootScope', '$location', '$routeParams', 'UserArticle', 'UserDraft', 'User','$analytics', '$window', 'Article', 'UserRecommendation', 'UserComment', 'UserLink',
+    function ($scope, $rootScope, $location, $routeParams, UserArticle, UserDraft, User, $analytics, $window, Article, UserRecommendation, UserComment, UserLink) {
 
     User.get({'userId': $routeParams.userId}, function(resource) {
       /* jshint camelcase: false */
@@ -10,15 +10,17 @@ angular.module('webClientApp')
       $rootScope.page.image = resource.cover_url;
       $rootScope.page.publishedTime = resource.created_at;
       $rootScope.page.description = resource.bio;
-      $scope.user = resource;
+      $scope.profile = resource;
     });
+
+    $rootScope.forceBar = true;
 
     $scope.editArticle = function (articleId) {
       $location.path('/articles/' + articleId + '/edit');
     };
 
     $scope.editProfile = function () {
-      $location.path('/profiles/' + $rootScope.currentUser.id + '/edit');
+      $location.path('/profiles/' + $rootScope.user.id + '/edit');
     };
 
     var deleteSuccess = function () {
@@ -56,6 +58,8 @@ angular.module('webClientApp')
 
     $scope.loadRecommendations = function() {
       $scope.activeTab = 'recommendations';
+      $scope.articles = [{ loading: true }, { loading: true },
+          { loading: true }];
       var articles = [];
       UserRecommendation.query({'userId': $routeParams.userId}, function(recommendations) {
         angular.forEach(recommendations, function (recommendation) {
@@ -67,6 +71,8 @@ angular.module('webClientApp')
 
     $scope.loadDiscussions = function() {
       $scope.activeTab = 'discussions';
+      $scope.articles = [{ loading: true }, { loading: true },
+            { loading: true }];
       var articles = [];
       UserComment.query({'userId': $routeParams.userId}, function(comments) {
         angular.forEach(comments, function (comment) {
@@ -85,11 +91,13 @@ angular.module('webClientApp')
     };
 
     $scope.loadArticles = function() {
+      $scope.articles = [{ loading: true }, { loading: true },
+            { loading: true }];
       $scope.activeTab = 'published';
       // Only get drafts if the current profile being viewed and the logged in user
       // are the same person.
-      if (($rootScope.currentUser &&
-           $rootScope.currentUser.id === parseInt($routeParams.userId))) {
+      if (($rootScope.user &&
+           $rootScope.user.id === parseInt($routeParams.userId))) {
         $scope.drafts = UserDraft.query({});
       }
 
@@ -98,6 +106,18 @@ angular.module('webClientApp')
       });
     };
 
+
+    $scope.loadLinks = function() {
+      UserLink.query({'userId': $routeParams.userId}, function (links) {
+        $scope.links = links;
+      });
+    };
+
+    $scope.getCardColor = function(color) {
+      return color || '#C0C0C0';
+    };
+    
+    $scope.loadLinks();
     $scope.loadArticles();
 
   }]);
