@@ -4,8 +4,6 @@ angular.module('webClientApp')
   .controller('SignupCtrl', ['$scope', '$location', '$routeParams', '$analytics', 'SignupService',
       function ($scope, $location, $routeParams, $analytics, SignupService) {
 
-    var isSignupPage = $location.path() === '/signup';
-    $scope.user = {};
     $scope.error = null;
     $scope.errorMessages = {};
 
@@ -14,13 +12,12 @@ angular.module('webClientApp')
     };
 
     var success = function() {
-      // TODO (HammamSamara) redirect to a view that requests email confirmation
       $analytics.eventTrack('Register Success', {
         category: 'User'
       });
-      if (isSignupPage) {
-        $location.path($routeParams.prev || '/');
-      }
+
+      $scope.flash = 'تم إرسال رسالة إلى بريدك الإلكتروني لتفعيل حسابك.';
+      $scope.flashNote = '(تأكد من مجلد السبام)';
     };
 
     var error = function(response) {
