@@ -8,6 +8,9 @@ angular.module('webClientApp')
       var ArticleCommentResource = $resource(
         baseUrl + 'articles/:articleId/comments/:commentId', {
           articleId: '@articleId'
+        }, {
+          get: {cache: true},
+          query: {cache: true, isArray: true}
         });
 
       return ArticleCommentResource;
@@ -17,8 +20,11 @@ angular.module('webClientApp')
       function ($resource, API_HOST) {
 
       var baseUrl = '//' + API_HOST + '/api/v1/';
-      var UserCommentResource = $resource(
-          baseUrl + 'users/:userId/comments');
+      var UserCommentResource = $resource(baseUrl + 'users/:userId/comments',
+        {}, {
+          get: {cache: true},
+          query: {cache: true, isArray: true}
+        });
 
       return {
         query: UserCommentResource.query
