@@ -8,6 +8,9 @@ angular.module('webClientApp')
       var ArticleRecommendationResource = $resource(
         baseUrl + 'articles/:articleId/recommendations/:recommendationId', {
           articleId: '@articleId'
+        }, {
+          get: {cache: true},
+          query: {cache: true, isArray: true}
         });
 
       return {
@@ -22,7 +25,10 @@ angular.module('webClientApp')
 
       var baseUrl = '//' + API_HOST + '/api/v1/';
       var UserRecommendationResource = $resource(
-          baseUrl + 'users/:userId/recommendations');
+        baseUrl + 'users/:userId/recommendations', {}, {
+          get: {cache: true},
+          query: {cache: true, isArray: true}
+        });
 
       return {
         query: UserRecommendationResource.query
